pyproject.toml
[project]
name = "telegram-terminal-bot"
version = "1.0.0"
description = "Control remote terminal sessions via Telegram bot"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=20.7,<21.0",
    "python-dotenv>=1.0.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "PT", "RET", "SIM"]
ignore = ["S101", "S603", "S607", "S110", "SIM105"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S105", "S106", "S108"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.11",
]
